Move GraphQLServiceContext into plugins, move ApolloConfig stuff to config.ts
