+++ packages/server/src/types.ts
@@ -9,14 +9,13 @@ import type {
 import type { IMocks } from '@graphql-tools/mock';
 import type { IExecutableSchemaDefinition } from '@graphql-tools/schema';
 import type {
-  ApolloConfig,
   GraphQLExecutor,
-  ApolloConfigInput,
   ApolloServerPlugin,
   BaseContext,
   GraphQLResponse,
   GraphQLRequestContext,
 } from './externalTypes';
+import type { ApolloConfig, ApolloConfigInput } from './config';
 import type { Logger } from '@apollo/utils.logger';
 
 import type { KeyValueCache } from 'apollo-server-caching';
