+++ packages/server/src/utils/schemaManager.ts
@@ -1,10 +1,10 @@
 import type { GraphQLSchema } from 'graphql';
+import type { Logger } from '@apollo/utils.logger';
+import type { ApolloConfig } from '../config';
 import type {
-  ApolloConfig,
   GraphQLExecutor,
   GraphQLSchemaContext,
 } from '../externalTypes';
-import type { Logger } from '@apollo/utils.logger';
 import type { GatewayInterface, Unsubscriber } from '../types';
 import type { SchemaDerivedData } from '../ApolloServer';
